% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtydistros.R
\name{fr.gamma}
\alias{fr.gamma}
\title{Gamma Frailty Distribution}
\usage{
fr.gamma(k, s, theta, what = "logLT")
}
\arguments{
\item{k}{Integer; derivative order.}

\item{s}{Numeric; argument of the Laplace transform.}

\item{theta}{Numeric > 0; heterogeneity parameter.}

\item{what}{Character; either \code{"logLT"} for \eqn{\log[ (-1)^k L^{{(k)}}(s) ]} or \code{"tau"} the Kendall's Tau.}
}
\value{
Numeric value of log-Laplace transform or tau.
}
\description{
Computes the log-Laplace transform or Kendall’s tau for the Gamma frailty model.
}
\keyword{internal}
