% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtydistros.R
\name{fr.ingau}
\alias{fr.ingau}
\title{Inverse Gaussian Frailty Distribution}
\usage{
fr.ingau(k, s, theta, what = "logLT")
}
\arguments{
\item{k}{Integer; derivative order.}

\item{s}{Numeric > 0; Laplace argument.}

\item{theta}{Numeric > 0; heterogeneity parameter.}

\item{what}{Character; either \code{"logLT"} for \eqn{\log[ (-1)^k L^{{(k)}}(s) ]} or \code{"tau"} the Kendall's Tau.}
}
\value{
Numeric value or message if tau is unstable.
}
\description{
Computes the log-Laplace transform or Kendall’s tau for the Inverse Gaussian frailty.
}
\keyword{internal}
