% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtydistros.R
\name{fr.lognormal}
\alias{fr.lognormal}
\title{Lognormal Frailty Distribution}
\usage{
fr.lognormal(k, s, sigma2, what = "logLT")
}
\arguments{
\item{k}{Order of derivative.}

\item{s}{Laplace argument.}

\item{sigma2}{Variance parameter (positive).}

\item{what}{Character; either \code{"logLT"} for \eqn{\log[ (-1)^k L^{{(k)}}(s) ]} or \code{"tau"} the Kendall's Tau.}
}
\value{
Numeric approximation of log-LT or Kendall's tau.
}
\description{
Computes the log-Laplace transform or Kendall's tau for the Lognormal frailty.
}
\keyword{internal}
