% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtydistros.R
\name{fr.none}
\alias{fr.none}
\title{No Frailty Distribution}
\usage{
fr.none(s, what = "logLT")
}
\arguments{
\item{s}{Argument of the Laplace transform.}

\item{what}{Character; either \code{"logLT"} for \eqn{\log[ (-1)^k L^{{(k)}}(s) ]} or \code{"tau"} the Kendall's Tau.}
}
\value{
Numeric value. \code{-s} if \code{what = "logLT"}, otherwise \code{NULL}.
}
\description{
Computes the Laplace transform for a model with no frailty.
}
\keyword{internal}
