% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtydistros.R
\name{fr.possta}
\alias{fr.possta}
\title{Positive Stable Frailty Distribution}
\usage{
fr.possta(k, s, nu, Omega, what = "logLT", correct)
}
\arguments{
\item{k}{Derivative order.}

\item{s}{Laplace argument.}

\item{nu}{Stability parameter in (0,1).}

\item{Omega}{Omega matrix of coefficients.}

\item{what}{Character; either \code{"logLT"} for \eqn{\log[ (-1)^k L^{{(k)}}(s) ]} or \code{"tau"} the Kendall's Tau.}

\item{correct}{Correction factor for large clusters.}
}
\value{
Numeric value.
}
\description{
Computes log-Laplace transform or tau for Positive Stable frailty.
}
\keyword{internal}
