% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtydistros.R
\name{J}
\alias{J}
\title{Polynomial Omega Sum for Positive Stable}
\usage{
J(k, s, nu, Omega, correct)
}
\arguments{
\item{k}{Order of derivative.}

\item{s}{Argument for Laplace.}

\item{nu}{Stability parameter.}

\item{Omega}{Matrix of omega coefficients.}

\item{correct}{Correction factor for large clusters.}
}
\value{
Numeric value of polynomial sum.
}
\description{
Computes helper sum over Omega matrix for Positive Stable frailty.
}
\keyword{internal}
